Create Separate SysConfig.h files for Surface and Rover
The Suface and Rover each follow the OpenRov2 software structure.  However, each
of these will use different modules.  The #defines in SysConfig.h allow a developer
to use pre-processor macros to defind of these two configurations will be conpiled.
